+++ hyperdbg/hyperkd/code/driver/Ioctl.c
@@ -1286,7 +1286,7 @@ DrvDispatchIoControl(PDEVICE_OBJECT DeviceObject, PIRP Irp)
             //
             // Perform the HyperTrace operation
             //
-            HyperTracePerformOperation(HyperTraceOperationRequest, TRUE);
+            HyperTraceLbrPerformOperation(HyperTraceOperationRequest, TRUE);
 
             Irp->IoStatus.Information = SIZEOF_HYPERTRACE_OPERATION_PACKETS;
             Status                    = STATUS_SUCCESS;