+++ hyperdbg/hypertrace/code/api/PtApi.c
@@ -1,0 +1,21 @@
+/**
+ * @file PtApi.c
+ * @author
+ * @brief Tracing routines for HyperTrace module (Intel Processor Trace)
+ * @details
+ * @version 0.19
+ * @date 2026-04-25
+ *
+ * @copyright This project is released under the GNU Public License v3.
+ */
+#include "pch.h"
+
+/**
+ * @brief Example of performing PT trace
+ *
+ * @return BOOLEAN
+ */
+VOID
+HyperTracePtExample()
+{
+}