+++ hyperdbg/hypertrace/code/pt/Pt.c
@@ -1,0 +1,21 @@
+/**
+ * @file Pt.c
+ * @author
+ * @brief Processor Trace (PT) tracing implementation for HyperTrace module
+ * @details
+ * @version 0.19
+ * @date 2026-04-25
+ *
+ * @copyright This project is released under the GNU Public License v3.
+ */
+#include "pch.h"
+
+/**
+ * @brief Processor Trace (PT) test function
+ *
+ * @return VOID
+ */
+VOID
+PtTest()
+{
+}