+++ hyperdbg/hypertrace/header/api/PtApi.h
@@ -1,0 +1,18 @@
+/**
+ * @file PtApi.h
+ * @author
+ * @brief Header for PT tracing routines for HyperTrace module (Intel Processor Trace)
+ * @details
+ * @version 0.19
+ * @date 2026-04-25
+ *
+ * @copyright This project is released under the GNU Public License v3.
+ */
+#pragma once
+
+//////////////////////////////////////////////////
+//			    	   Globals	    			//
+//////////////////////////////////////////////////
+
+VOID
+HyperTracePtExample();